#9 feat: SearchBar 컴포넌트 스타일 파일 분리 및 SearchResult 컴포넌트 생성
