+++ src/Components/SearchBar/SearchResult/SearchResult.jsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+const SearchResult = (props) => <div>result</div>;
+
+export default SearchResult;
