+++ src/Components/Tag/Tag.style.js
@@ -11,6 +11,7 @@ const InfoTag = styled.div`
   border: 1px solid #d9d9d9;
   box-sizing: border-box;
   border-radius: 100px;
+  color: black;
 
   ${(props) =>
     props.role === 'category' &&
@@ -21,7 +22,8 @@ const InfoTag = styled.div`
   ${(props) =>
     props.checked === true &&
     css`
-      background: #bdbdbd;
+      color: white;
+      background: #616161;
     `}
 `;
 
@@ -31,7 +33,6 @@ const TagText = styled.span`
   font-weight: normal;
   font-size: 12px;
   line-height: 20px;
-  color: rgba(0, 0, 0, 0.85);
 `;
 
 export const style = {
